cglbdef, cmain
